# libbeat: use optimized version of io.ReadAll to read  _bulk responses