## Proposed commit message

This change introduces an optimization that uses bytes.Buffer+io.Copy to read the response body in calls to the _bulk endpoint. Benchmark results and rationale were already discussed at https://github.com/elastic/beats/issues/36151 and a function optimized for reading http responses was added to the elastic-agent-libs at https://github.com/elastic/elastic-agent-libs/pull/183.

## Checklist

- [x] My code follows the style guidelines of this project
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation
- [ ] I have made corresponding change to the default configuration files
- [ ] I have added tests that prove my fix is effective or that my feature works
- [ ] I have added an entry in `CHANGELOG.next.asciidoc` or `CHANGELOG-developer.next.asciidoc`.

## How to test this PR locally

Any calls that use the Bulk api, there is existing testing coverage for this method as well.

## Related issues

- For https://github.com/elastic/beats/issues/36151
- Relates https://github.com/elastic/elastic-agent-libs/pull/183