@mauri870 2024-09-17
The old version seems to be faster, but the benchmark is using a simple hello world response, I have to change the benchmark to try different response sizes to be sure. I agree that it is likely that the buffer reuse is still faster than this approach.

```
goos: linux
goarch: amd64
pkg: github.com/elastic/beats/v7/libbeat/esleg/eslegclient
cpu: Intel(R) Xeon(R) CPU E5-2697A v4 @ 2.60GHz
                   │   old.txt   │               new.txt               │
                   │   sec/op    │   sec/op     vs base                │
ExecHTTPRequest-32   6.263µ ± 2%   7.082µ ± 0%  +13.08% (p=0.000 n=10)

                   │   old.txt    │               new.txt                │
                   │     B/op     │     B/op      vs base                │
ExecHTTPRequest-32   5.539Ki ± 0%   7.086Ki ± 0%  +27.93% (p=0.000 n=10)

                   │  old.txt   │              new.txt               │
                   │ allocs/op  │ allocs/op   vs base                │
ExecHTTPRequest-32   25.00 ± 0%   28.00 ± 0%  +12.00% (p=0.000 n=10)
```