@mauri870 2024-09-18
@rdner I ran some custom benchmarks and for this particular case the bytes.Buffer reuse is still faster. I'm going to go ahead and close this one, I will open a PR shortly with some improvements to the ExecHTTPRequest benchmark.